polypus
Module algorithms

Structs

AlgorithmArgs

Base arguments for all algorithms.

AlgorithmDifferentialEvolution

Algorithm to optimize a QAOA using differential evolution.

AlgorithmDifferentialEvolutionArgs

Arguments for the Differential Evolution algorithm.

AlgorithmDistributeByShots

Algorithm to distribute the exectuion of a quantum circuit by shots.

AlgorithmSingleRun

Single run of a quantum circuit.

Traits

AlgorithmTrait

Algorithm Trait Definition This trait defines the interface for all algorithms in Polypus.

Functions

differential_evolution

Function to train a QAOA using differnetial evolution called from Python.

raise_qpus

Function to raise QPUs in an infrastructure called from Python.

run_quantum_circuit

Function to run a quantum circuit called from Python.