polypus::algorithms
Function differential_evolution

pub fn differential_evolution<'py>(
    qc: Bound<'py, PyAny>,
    shots: Option<Bound<'py, PyInt>>,
    n_qpus: Option<u32>,
    expectation_function: Option<Bound<'py, PyAny>>,
    generations: Option<usize>,
    population_size: Option<usize>,
    dimensions: Option<usize>,
    infraestructure: Option<String>,
    nodes: Option<u32>,
    id: Option<String>,
    tolerance: Option<f64>,
) -> PyObject

Function to train a QAOA using differnetial evolution called from Python.

Arguments

qc: Quantum Circuit.

shots: Number of shots to run the quantum circuit.

n_qpus: Number of QPUs to use.

expectation_function: Expectation function to use. This function must be defined in Python.

generations: Number of max generations to run the differential evolution.

population_size: Size of the population (number of individuals).

dimensions: Number of dimensions of the problem to solve.

infraestructure: Hardware infraestructure to use. Use “local” if no specific infraestructure is available.

nodes: Number of nodes to use in the infrastructure. If infraestructure is “local”, this parameter is ignored.

id: Identifier for the algoirthm run. If multiples instances of polypus are used for different runs, use different ids to handle different log and temporary files.

tolerance: Tolerance for the convergence of the algorithm (early stop).

Returns

PyObject: Best individual parameters found by the differential evolution algorithm.

Examples

import polypus
# def bitstring_to_obj(bitstring):
best_parameters = polypus.differential_evolution(
  qc=qc, 
  shots=N_SHOTS, 
  n_qpus = N_QPUS, 
  expectation_function=bitstring_to_obj, 
  generations=MAX_GENERATIONS, 
  population_size=POPULATION_SIZE, 
  dimensions=2*layers, 
  infraestructure="qmio", 
  id=id,
  tolerance=TOL,
  nodes=N)