polypus::algorithms
Function run_quantum_circuit

pub fn run_quantum_circuit<'py>(
    qc: Bound<'py, PyAny>,
    shots: Option<Bound<'py, PyInt>>,
    infraestructure: Option<String>,
    n_qpus: Option<u32>,
) -> PyObject

Function to run a quantum circuit called from Python.

Arguments

qc: Quantum Circuit.

shots: Number of shots to run the quantum circuit.

infraestructure: Hardware infraestructure to use. Use “local” if no specific infraestructure is available.

n_qpus: Number of QPUs to use.

Returns

PyObject: Counts of the quantum circuit execution.

Examples

We can run a single quantum circuit:

import polypus
result = polypus.run_quantum_circuit(qft_circuit, shots=1000, infraestructure="local")

We can also run that circuit and distribute the execution accross multiple QPUs:

import polypus
result = polypus.run_quantum_circuit(qft_circuit, shots=1000, infraestructure="local",  n_qpus=10)