polypus::algorithms
Trait AlgorithmTrait

pub trait AlgorithmTrait {
    type Args<'py>;
    type AlgorithmReturnType: Display;

    // Required methods
    fn run<'py>(&self, args: Self::Args<'py>) -> Self::AlgorithmReturnType;
    fn name(&self) -> String;
    fn description(&self) -> String;
}

Algorithm Trait Definition This trait defines the interface for all algorithms in Polypus.

Args: Type of arguments the algorithm accepts.

AlgorithmReturnType: Type of the return value of the algorithm.

run: Method to execute the algorithm with the provided arguments.

name: Method to get the name of the algorithm.

description: Method to get the description of the algorithm.

Required Associated Types

type Args<'py>

type AlgorithmReturnType: Display

Required Methods

fn run<'py>(&self, args: Self::Args<'py>) -> Self::AlgorithmReturnType

fn name(&self) -> String

fn description(&self) -> String

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.